feat(deployment): populate configure screen from template query param
Resolve a templateId query param into the Configure screen's starting SDL,
matching the legacy flow: hardcoded templates (e.g. hello-world) are matched
by code and carry their SDL inline, while other ids are fetched as public
gallery templates via a new usePublicTemplate hook. Split ConfigureDeployment
into a thin resolver (handles param/atom resolution, loading, fetch errors)
and ConfigureDeploymentForm (the existing synchronous form, now seeded by an
initialSdl prop) so the form keeps initializing from a single source.
