+++ apps/deploy-web/src/queries/queryKeys.ts
@@ -17,6 +17,7 @@ export class QueryKeys {
   static getValidatorsKey = () => ["VALIDATORS"];
   static getProposalsKey = () => ["PROPOSALS"];
   static getTemplateKey = (id: string) => ["SDL_TEMPLATES", id];
+  static getPublicTemplateKey = (id: string) => ["PUBLIC_TEMPLATE", id];
   static getUserTemplatesKey = (username: string) => ["USER_TEMPLATES", username];
   static getUserFavoriteTemplatesKey = (userId: string) => ["USER_FAVORITES_TEMPLATES", userId];
   // Deploy
